Fix logic for building the <dl> in SubagesWithSummaries
This macro previously had bugs that caused its output to sometimes
include unescaped HTML tags, breaking content. I've revamped the
logic of building the list to be more intelligent about this and to
handle generation of the HTML more efficiently as well.

THIS PATCH REQUIRES THE PATCH TO FIX LandingPageListSubpages TO GO IN
THE SAME PRODUCTION PUSH. SubpagesWithSummaries is used by the latter,
and if the changes for either one go onto production independently,
pages *will* break.